Teste

+++ app/src/main/java/com/projeto/bookfast/bookfast/dominio/livro/gui/TelaCadastrarLivroAdministrador.java
@@ -46,7 +46,7 @@ public class TelaCadastrarLivroAdministrador extends Activity {
                 ViewGroup group = (ViewGroup) findViewById(R.id.raizCadastroLivro);
                 LimparTela limparTela = new LimparTela();
                 if (!validarCampo.vefificaCadastroLivro(editIsbn, editNome, editGenero, editAutor, editEdicao, editAno, editQuantidadeTotal, editQuantidadeAlugada)) {
-                    long isbn = Integer.parseInt(editIsbn.getText().toString());
+                    long isbn = Long.parseLong(editIsbn.getText().toString());
                     int edicao = Integer.parseInt(editEdicao.getText().toString());
                     int ano = Integer.parseInt(editAno.getText().toString());
                     int quantidadeTotal = Integer.parseInt(editQuantidadeTotal.getText().toString());
